session check

+++ src/useApiServices.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { persistor } from "./store";
 import { clearUserInfo } from "./store/userInfoSlice";
 import { clearNotificationCount } from "./store/notificationCountSlice";
@@ -8,11 +8,18 @@ import { clearThemeColor } from "./store/themeColorSlice";
 import { clearJobRoleSelect } from "./store/selectJobeRoleSlice";
 import { clearCompanySelect } from "./store/selectCompanySlice";
 import { clearEmployeeformFilled } from "./store/EmployeeFormSlice";
+import { clearSession, setSession } from "./store/SessionSlice";
+import { useRef } from "react";
+import { showToast } from "./main/ToastManager";
 
 const useApiServices = () => {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-
+  const sessionExpired = useSelector((state) => state.session.session);
+  const unauthorizedCalled = useRef(false);
+  if (typeof window.sessionExpired === "undefined") {
+    window.sessionExpired = false;
+  }
   const BASE_URL = process.env.REACT_APP_API_URL;
 
   const getAuthToken = () => {
@@ -25,14 +32,39 @@ const useApiServices = () => {
     Authorization: `Bearer ${getAuthToken()}`,
   });
 
-  const handleUnauthorized = async () => {
+  // const handleUnauthorized = async () => {
+  //   // if (sessionExpired) return;
+  //   console.warn("Session expired");
+  //   localStorage.clear();
+  //   dispatch(clearUserInfo());
+  //   dispatch(clearNotificationCount());
+  //   dispatch(clearThemeColor());
+  //   dispatch(clearJobRoleSelect());
+  //   dispatch(clearCompanySelect());
+  //   dispatch(clearEmployeeformFilled());
+  //   persistor.pause();
+  //   await persistor.purge();
+  //   persistor.persist();
+  //   navigate("/login");
+  // };
+
+  const handleUnauthorized = async (errorMessage) => {
+    if (unauthorizedCalled.current || window.sessionExpired) return;
+    unauthorizedCalled.current = true;
+    window.sessionExpired = true;
+    showToast(errorMessage, "error");
     localStorage.clear();
+    // navigate("/login");
+    await new Promise((resolve) => setTimeout(resolve, 500));
+
     dispatch(clearUserInfo());
     dispatch(clearNotificationCount());
     dispatch(clearThemeColor());
     dispatch(clearJobRoleSelect());
     dispatch(clearCompanySelect());
     dispatch(clearEmployeeformFilled());
+    dispatch(clearSession());
+    window.sessionExpired = false;
     persistor.pause();
     await persistor.purge();
     persistor.persist();
@@ -40,6 +72,10 @@ const useApiServices = () => {
   };
 
   const GetCall = async (endpoint, params = {}, headers = {}) => {
+    console.log(" before get call");
+    if (sessionExpired || window.sessionExpired) return;
+    console.log("get call");
+
     try {
       const response = await axios.get(`${BASE_URL}${endpoint}`, {
         params,
@@ -49,15 +85,25 @@ const useApiServices = () => {
         },
       });
       if (response?.data?.status === 5000) {
-        await handleUnauthorized();
+        dispatch(setSession(true));
+        console.log(response);
+        await handleUnauthorized(response?.data?.message);
+        // await handleUnauthorized();
+        return;
       }
+      console.log(("response", response));
       return response;
     } catch (error) {
       console.error("GET request error:", error);
+      return error.response;
     }
   };
 
   const PostCall = async (endpoint, body = {}, headers = {}) => {
+    console.log(" before post call");
+    if (sessionExpired || window.sessionExpired) return;
+    console.log("poast call");
+
     try {
       const response = await axios.post(`${BASE_URL}${endpoint}`, body, {
         headers: {
@@ -66,8 +112,12 @@ const useApiServices = () => {
         },
       });
       if (response?.data?.status === 5000) {
-        await handleUnauthorized();
+        dispatch(setSession(true));
+        console.log(response);
+        await handleUnauthorized(response?.data?.message);
+        return;
       }
+      console.log(("response", response));
       return response;
     } catch (error) {
       console.error("POST request error:", error);
